Fixes for bug 324655.
There are two ways to set the scrolling.  ResourcesTab currently uses the
setMinSize method, but with my dynamically built composite this fails until it
gets updated.  If, however, I switch the method to have the ResourcesTab set
the size on the control itself (from the computed size of the control), PBS is
happy.  The only other thing the PBS tab needs to do is turn off the
ScrolledComposite expansion (left on because the other RMs seem to need it).

I've checked the look of the current RM tabs and they seem to be unchanged with
this modification.