Patch for bug 324654.
The Launch Tab for PBS now displays the list of machine queues available as a combo box, values derived from the IPQueue elements sent during RM model definition.

The auxiliary method for populating the box is improperly encapsulated (it belongs in the Wizard Page), but this would have required an API change.

Also, "destination" must be a fixed item in the PBS client's attribute name vocabulary for this to work.