JAXB control wizard page disables everything until an XML config resource is selected.  When it is selected, it is immediately unmarshaled into the data object.  The ResourceManager now gets this object from the service provider (configuration) instead of deserializing it internally.  When the selection is made, the connection fields are enabled.  If there are site defaults, the remote target path appears instead of blank text in the "service path" text box (e.g., "/usr/local/bin/qsub").
Modified the org.eclipse.ptp.remote.ui.IRemoteUIConnectionManager interface to take an overloaded method, public IRemoteConnection newConnection(Shell shell, Map<String, String> defaultAttr), and then to modify the remote tools UI implementation to allow for a TargetElement wrapping the TargetTypeElement and adding the two attributes, ConfigFactory.ATTR_CONNECTION_ADDRESS and ConfigFactory.ATTR_CONNECTION_PORT, with the values coming from the <site> definition in the XML; for Local and RSE UI managers, the new overloaded method do the same thing as the original.

IControlMonitorRMConfiguration provides methods for getting and setting separate paths, invocation options and connection names on both control and monitor.  In the latter case, each wizard page has a wrapper method which stores the name according to type (control or monitor).

The wizard for control is used to establish the main connection.  The monitoring wizard has a checkbox for sharing the control connection; if this is checked, the wizard disables the widgets associated with connection selection, and saves the name of the control connection as the monitoring connection as well.  If not, one can use the wizard to establish a different connection.

I have not yet figured out how to designate this connection HTTP rather than SSH.  I would guess we need a new remote service to the list of available services (e.g, local, remotetools, rse, http).

In any case, a single remote connection manager looks like it is capable of handling different connections, provided they are named differently.  I need to put in place some mechanism whereby the user cannot reuse the control connection name.