Changed the way configurations for JAXB RMs are handled.
A JAXB config now is stored directly in the RM Config as XML.  This avoids the problem of the user removing it from the workspace or the url changing when plugins are updated.

This now means the the JAXB config actually defines the RM instance throughout its lifecycle, and once it is instantiated the first time this definition cannot be changed.