Modifications to support forced reloading of XML from URL. Added preference management and page. The configuration now checks these conditions when constructing the data tree:
if the config XML string is null, or force is set to true, it uses the URL originally passed in in the wizard; otherwise, it uses the cached URL.

In the case that the force option is turned on and the file and the stored URL is missing, these warnings will appear:

Start:  Failed to start resource manager:  Reason:  Failed to create ResourceManager data tree.
Open of Launch Tab:

The configuration file for this resource manager is either missing or invalid; the resource manager has become unusuable and should be discarded.

Failed to create ResourceManager data tree.