Bug fix 354001
The issue here was establishing the correct initial environment. The inclusion of a widget value in the environment depends on, among other things, its visibility; however, the visible property of a widget cannot be forced unless its ancestors are visible as well.  When a configuration is first created, none of the tabs have been rendered visible, so any shared values for the first tab are not going to be set until those other tabs are actually exposed.

The fix for this was to add an "initialized" property to the configuration itself; if this is false, then the visible property of the widgets is bypassed.  This is essentially equivalent to turning on visible for all widgets by default.