Removing "eclipse" from gitignore
Sorry this caused probably most problems with adding files.
I had added it because some of the jaxb projects build the .class files
directly in "eclipse" not in "bin" as other projects. But this is only a
minor problem but adding "eclipse" causes all source files to be
ignored, because "org/eclipse/...." includes "eclipse".