Fix display bug in new sync wizard on Windows
The groups in the project type couldn't be opened on Windows. Somehow
caused by autotools (bug only appeared with autotools first in list).
This change makes the first group and not the first element in the first
group selected by default. This also makes the list a bit easier to read
if the first group has many items (no scroll bars).